requirements.txt
# Core data
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# Database
psycopg2-binary>=2.9.0

# ML
scikit-learn>=1.3.0
mlflow>=2.8.0

# Orchestration (Airflow - installed separately via Docker)
# apache-airflow is managed by the docker-compose image

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# UI
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0
